make configuration and add address

+++ src/main/java/org/zerhusen/repository/DistrictRepository.java
@@ -3,6 +3,6 @@ package org.zerhusen.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.zerhusen.model.District;
 
-public interface DistrictRepository extends JpaRepository<District, Long> {
+public interface DistrictRepository extends JpaRepository<District, Integer> {
 	
 }
